dash/requirements.txt
databricks-sql-connector==4.0.0
databricks-sdk==0.45.0
dash==3.0.0
posit-sdk==0.9.0
